text_pil: prevent warning from pil imaging library

+++ pymt/core/text/text_pil.py
@@ -40,8 +40,8 @@ class LabelPIL(LabelBase):
         self._pil_draw = ImageDraw.Draw(self._pil_im)
 
     def _render_text(self, text, x, y):
-        color = tuple(map(lambda x: x * 255, self.options['color']))
-        self._pil_draw.text((x, y), text, font=self._select_font(), fill=color)
+        color = tuple(map(lambda x: int(x * 255), self.options['color']))
+        self._pil_draw.text((int(x), int(y)), text, font=self._select_font(), fill=color)
 
     def _render_end(self):
         data = pymt.ImageData(self.width, self.height,
